pyproject.toml
[project]
name = "scrapka"
version = "0.1.0"
description = "Google Maps Scraper with Camoufox"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "camoufox[geoip]>=0.4.0",
    "playwright>=1.40.0",
    "openpyxl>=3.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "ruff>=0.15.1",
    "python-dotenv>=1.2.1",
]

[tool.ruff]
line-length = 150
indent-width = 2
